MULE-15586: Add a timeout to lookup function * Fir for classcastexception on some scenarios

+++ core/src/main/java/org/mule/runtime/core/internal/el/function/LookupFunction.java
@@ -126,7 +126,7 @@ public class LookupFunction implements ExpressionFunction {
                     if (schedulerService.isCurrentThreadForCpuWork()) {
                       return 2000;
                     } else {
-                      return SECONDS.toMillis(60);
+                      return (int) SECONDS.toMillis(60);
                     }
                   }));
   }
